+++ db/upgrade.php
@@ -526,5 +526,41 @@ function xmldb_theme_boost_union_upgrade($oldversion) {
         upgrade_plugin_savepoint(true, 2024100716, 'theme', 'boost_union');
     }
 
+    if ($oldversion < 2025041400) {
+
+        // Remove the mediumcontentmaxwidth setting from Boost Union.
+        unset_config('mediumcontentmaxwidth', 'theme_boost_union');
+
+        // Remove the activitiestintenabled setting from Boost Union.
+        unset_config('activitiestintenabled', 'theme_boost_union');
+
+        // Remove the activityiconcolorfidelity setting from Boost Union.
+        unset_config('activityiconcolorfidelity', 'theme_boost_union');
+
+        // Boost_union savepoint reached.
+        upgrade_plugin_savepoint(true, 2025041400, 'theme', 'boost_union');
+    }
+
+    if ($oldversion < 2025041401) {
+
+        // Set the guestroleupgradedfrompre500 marker to be used by the settings_update_guestrole.php script.
+        set_config('guestroleupgradedfrompre500', 1, 'theme_boost_union');
+
+        // Boost_union savepoint reached.
+        upgrade_plugin_savepoint(true, 2025041401, 'theme', 'boost_union');
+    }
+
+    if ($oldversion < 2025041402) {
+
+        // Remove the THEME_BOOST_UNION_SETTING_SLIDER_ANIMATIONTYPE_NONE option from the slideranimation setting.
+        $oldconfig = get_config('theme_boost_union', 'slideranimation');
+        if ($oldconfig == 0) {
+            set_config('slideranimation', THEME_BOOST_UNION_SETTING_SLIDER_ANIMATIONTYPE_SLIDE, 'theme_boost_union');
+        }
+
+        // Boost_union savepoint reached.
+        upgrade_plugin_savepoint(true, 2025041402, 'theme', 'boost_union');
+    }
+
     return true;
 }